Flesh out temporary directory and file-pathing more substantially
