+++ caiman/paths.py
@@ -4,6 +4,7 @@
 """
 
 import os
+import logging
 from typing import Tuple
 
 
@@ -25,6 +26,38 @@ def caiman_datadir() -> str:
 def caiman_datadir_exists() -> bool:
     return os.path.isdir(caiman_datadir())
 
+######
+# tempdir
+
+def get_tempdir() -> str:
+    """ Returns where CaImAn can store temporary files, such as memmap files. Controlled mainly by environment variables """
+    # CAIMAN_TEMP is used to control where temporary files live. 
+    # If unset, uses default of a temp folder under caiman_datadir()
+    # To get the old "store it where I am" behaviour, set CAIMAN_TEMP to a single dot.
+    # If you prefer to store it somewhere different, provide a full path to that location.
+    if 'CAIMAN_TEMP' in os.environ:
+        if os.path.isdir(os.environ['CAIMAN_TEMP']):
+            return os.environ['CAIMAN_TEMP']
+        else:
+            logging.warning(f"CAIMAN_TEMP is set to nonexistent directory {os.environment['CAIMAN_TEMP']}. Ignoring")
+    temp_under_data = os.path.join(caiman_datadir(), "temp")
+    if not os.path.isdir(temp_under_data):
+        logging.warning(f"Default temporary dir {temp_under_data} does not exist, creating")
+        os.makedirs(temp_under_data)
+    return temp_under_data
+
+def fn_relocated(fn:str) -> str:
+    """ If the provided filename does not contain any path elements, this returns what would be its absolute pathname
+        as located in get_tempdir(). Otherwise it just returns what it is passed.
+
+        The intent behind this is to ease having functions that explicitly mention pathnames have them go where they want,
+        but if all they think about is filenames, they go under CaImAn's notion of its temporary dir. This is under the
+        principle of "sensible defaults, but users can override them".
+    """
+    if str(os.path.basename(fn)) == str(fn): # No path stuff
+        return os.path.join(get_tempdir(), fn)
+    else:
+        return fn
 
 ######
 # memmap files
